Refs #124. Abstracted out the fixture factory test case generator
